art_gallery_portfolio/api/requirements.txt
Flask
python-dotenv
mysql-connector-python
gunicorn
Flask-Cors
flask-bcrypt